requirements.txt
Flask>=2.3.0
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
Werkzeug>=2.3.0

